Move Domain class to file, small exception changes
