+++ main_code.py
@@ -1,49 +1,18 @@
-import requests, datetime, os
-from datetime import timedelta
+import os
+from Domain import Domain
 from Mysql import create_connection, get_domains_list
-from mysql.connector import MySQLConnection
 from dotenv import load_dotenv
 
-class Domain():
-    def __init__(self,domain, exp_date) -> None:
-        self.domain = domain
-        if exp_date:
-            self.exp_date = exp_date
-    
-    def get_exp_dates(self):
-        try:
-            api_key = os.getenv(f"TEMP_API_KEY")
-            response = requests.get(f"https://www.whoisxmlapi.com/whoisserver/WhoisService?apiKey={api_key}&domainName={self.domain}&outputFormat=JSON")
-            main_responce_dict = response.json()
-            exp_date_time=main_responce_dict['WhoisRecord']['expiresDateNormalized']
-            print(exp_date_time)
-            self.exp_date = datetime.datetime.strptime(exp_date_time, '%Y-%m-%d %H:%M:%S UTC').date()
-            return self
-        except Exception as e:
-            print(f"Exception Thrown while calling API: {e}")
-        
-    
-    def compare_dates(self) -> timedelta:
-        now=datetime.datetime.now().date()
-        if 'exp_date' not in vars(self).keys(): 
-            self.get_exp_dates()
-        return self.exp_date - now
-    
-    def updateDateInDatabase(self, db: MySQLConnection):
-        try:
-            date = self.exp_date.strftime("\'%y-%m-%d\'")
-            cursor = db.cursor()
-            cursor.execute(f"update domains SET date = {date} WHERE domain = \'{self.domain}\';")
-            db.commit()
-        finally:
-            cursor.close()
-
 def main():
-    domains=[]
     load_dotenv()
+    domains=[]
     try:
-        db=create_connection()
-    
+        db=create_connection(
+            host = os.getenv("HOST_SQL"),
+            user = os.getenv("USER_SQL"),
+            password = os.getenv("PASSWORD_SQL"),
+            database = os.getenv("DB_SQL")
+        )
         for domain in get_domains_list(db): 
             if domain[2] == None:
                 dom = Domain(domain[1],None)
@@ -54,9 +23,11 @@ def main():
             domains.append(dom)
             if dom.compare_dates().days < 30: #Do Date Compare
                 print("This is email send function")
-
+    except Exception as e:
+        print(e) #Write more spesific exception and adopt logging methods
     finally:
-        db.close()
+        if db:
+            db.close()
 
 if __name__ == "__main__":
     main()
